Add functions to find limit and offset number token value (#31) (#32)
After enforcing the value to the statement, it would be useful to
know the limit and offset values, regardless of whether they were
enforced on the statement or already defined in the statement.

This brings more customizability to users if they use the
`getStatementClasses` function. They can create their own
`limit` function to return desired values.